Rename AlphabetStrip
