+++ src/lib/components/LexiconAlphabetStrip.svelte
